Changed entity to folder entities
